factor reviewDetails into singlereview
