--- frontend/src/components/ReviewDetails.css
@@ -1,23 +1,0 @@
-.add-comment-container {
-    margin-top: 20px;
-}
-
-textarea {
-    width: 100%;
-    padding: 8px;
-    margin-bottom: 10px;
-    resize: vertical;
-}
-
-button {
-    background-color: #007bff;
-    color: #fff;
-    padding: 8px 12px;
-    border: none;
-    cursor: pointer;
-    border-radius: 4px;
-}
-
-button:hover {
-    background-color: #0056b3;
-}
